Move observation activity guard to ActivityMediaService
