+++ src/main/kotlin/com/terraformation/backend/accelerator/db/ActivityMediaStore.kt
@@ -11,7 +11,6 @@ import com.terraformation.backend.db.accelerator.tables.pojos.ActivityMediaFiles
 import com.terraformation.backend.db.accelerator.tables.records.ActivitiesRecord
 import com.terraformation.backend.db.accelerator.tables.references.ACTIVITIES
 import com.terraformation.backend.db.accelerator.tables.references.ACTIVITY_MEDIA_FILES
-import com.terraformation.backend.db.accelerator.tables.references.ACTIVITY_OBSERVATIONS
 import com.terraformation.backend.db.asNonNullable
 import com.terraformation.backend.db.default_schema.FileId
 import com.terraformation.backend.db.default_schema.OrganizationId
@@ -192,10 +191,6 @@ class ActivityMediaStore(
     withLockedActivity(activityId) {
       ensureFileExists(activityId, fileId)
 
-      if (isObservation(activityId)) {
-        throw CannotDeleteObservationActivityMediaException(activityId)
-      }
-
       dslContext
           .deleteFrom(ACTIVITY_MEDIA_FILES)
           .where(ACTIVITY_MEDIA_FILES.FILE_ID.eq(fileId))
@@ -319,12 +314,6 @@ class ActivityMediaStore(
           .execute()
     }
   }
-
-  private fun isObservation(activityId: ActivityId): Boolean =
-      dslContext.fetchExists(
-          ACTIVITY_OBSERVATIONS,
-          ACTIVITY_OBSERVATIONS.ACTIVITY_ID.eq(activityId),
-      )
 
   private fun markActivityModified(activityId: ActivityId) {
     with(ACTIVITIES) {
